export and refactor
