+++ react-exercises/snake-v2/src/styles/newApp.css
